Upgrading dependencies to allow building on Python 3.9
With the current version numbers, running "docker-compose up --build" on privacy-api with Python 3.9 used in the privacy_api Dockerfile would result in a 9 to 10 minute wait for "pip install" to complete. This appears to be due an older spacy dependency; if I leave that at 2.2.3, it takes 9 to 10 minutes; at 2.3.7, it takes 80s.

Another issue this solves is that when running "pip install -e . --no-cache-dir" on this project with Python 3.9.x, the install would fail. Upgrading cython to the latest minor version fixed this problem. 

I also upgraded grpcio and protobuf to the minimum versions in order to resolve some dependency conflict errors that appeared when running "pip install -e . --no-cache-dir".