Quieting down some logging in recognizer_registry
The "Returning a total" message appears every time that the presidio PII analysis is run on a field, muddying up the log messages quite a bit. To allow privacy-api to control this, modified recognizer_registry to use a named logger and changed all the log calls to debug instead of info.